% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computationAndAnalysis.R
\name{analyzePopulationStructure}
\alias{analyzePopulationStructure}
\title{Analyze population structure}
\usage{
analyzePopulationStructure(
  vcf_data,
  pop_metadata,
  method = "pca",
  n_components = 2,
  min_var = 0.01
)
}
\arguments{
\item{vcf_data}{A CollapsedVCF object}

\item{pop_metadata}{Population metadata data.frame}

\item{method}{One of "pca"}

\item{n_components}{Number of components}

\item{min_var}{Minimum variance threshold for filtering variants}
}
\value{
List containing analysis results
}
\description{
A function that runs PCA analysis(can me modified for other analyses).
Output is ready for plotting various plot types.
Warning: When sample size is too small, you may receive the following Warning messages:
1: In MASS::cov.trob(data\link{, vars}) : Probable convergence failure, or similar
}
\examples{
# Grab filtered_data from previous load/filter functions

pca_results <- genRelateR::analyzePopulationStructure(
 filtered_data$vcf_data,
 filtered_data$pop_metadata,
 method = "pca"
)

}
