% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computationAndAnalysis.R
\name{convertGTtoNumeric}
\alias{convertGTtoNumeric}
\title{Function to validate and process genotype data}
\usage{
convertGTtoNumeric(gt_data)
}
\arguments{
\item{gt_data}{Genotype data results from extractGenotypeFromCollapsed}
}
\value{
Converted genotype data in a matrix
}
\description{
Function to validate and process genotype data
}
