% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computationAndAnalysis.R
\name{extractGenotypeFromCollapsed}
\alias{extractGenotypeFromCollapsed}
\title{Function to extract genotype data from CollapsedVCF}
\usage{
extractGenotypeFromCollapsed(collapsed_vcf)
}
\arguments{
\item{collapsed_vcf}{Vcf data in the collapsed form}
}
\value{
Genotype data from a collapsedVCF
}
\description{
Function to extract genotype data from CollapsedVCF
}
