% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadAndCleanData.R
\name{filterPopulation}
\alias{filterPopulation}
\title{Filter Population Data}
\usage{
filterPopulation(vcf_data, pop_file, population = NULL, super_pop = NULL)
}
\arguments{
\item{vcf_data}{A VCF object loaded using loadGeneticData}

\item{pop_file}{Path to population metadata file (tab-delimited with columns: Sample, Population, SuperPop)}

\item{population}{Vector of population codes to include (e.g., c("CEU", "YRI"))}

\item{super_pop}{Vector of super-population codes (e.g., c("EUR", "AFR"))}
}
\value{
Filtered VCF object
}
\description{
This function filters VCF data based on population metadata from the 1000 Genomes Project.
}
\examples{
# Load genetic data and metadata
genRelateR::loadGeneticDada(vcf_file)
pop_metadata <- "data/population_metadata.txt"

# Filter populations
populations <- c(  "CHB", "JPT", "CHS", "CDX", "KHV", "CEU", "TSI", "GBR",
"FIN", "IBS", "YRI", "LWK", "GWD", "MSL", "ESN","ASW", "ACB", "MXL", "PUR",
"CLM", "PEL", "GIH", "PJL", "BEB", "STU", "ITU")

filtered_data <- genRelateR::filterPopulation(genetic_data, pop_metadata, populations)

}
