% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computationAndAnalysis.R
\name{filterVariantsByVariance}
\alias{filterVariantsByVariance}
\title{Filter variants by variance}
\usage{
filterVariantsByVariance(geno_mat, min_var = 1e-10)
}
\arguments{
\item{geno_mat}{Numeric genotype matrix}

\item{min_var}{Minimum variance threshold}
}
\value{
Filtered genotype matrix
}
\description{
Filter variants by variance
}
