% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizationScripts.R
\name{plotPopulationPca}
\alias{plotPopulationPca}
\title{Plot Population Structure PCA}
\usage{
plotPopulationPca(
  analysis_results,
  pop_metadata,
  title = NULL,
  ellipses = TRUE,
  super_pop = FALSE
)
}
\arguments{
\item{analysis_results}{Results from analyzePopulationStructure() with method="pca"}

\item{title}{Plot title (optional)}

\item{ellipses}{Logical indicating whether to draw confidence ellipses (default: TRUE)}

\item{labels}{Logical indicating whether to show sample labels (default: FALSE)}
}
\value{
ggplot object
}
\description{
Creates a scatter plot of the first two principal components from population structure analysis
with populations colored distinctly and confidence ellipses.
}
\examples{
# Grab pca_results from previous analysis functions
pca_results <- genRelateR::analyzePopulationStructure(
 filtered_data$vcf_data,
 filtered_data$pop_metadata,
 method = "pca"
)

pca_plot <- genRelateR::plotPopulationPca(
  analysis_results = pca_results,
  filtered_data$pop_metadata,
  title = "Population Structure PCA",
  ellipses = TRUE,
  super_pop = TRUE
  )
print(pca_plot)

}
