% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rungenRelateR.R
\name{rungenRelateR}
\alias{rungenRelateR}
\title{Launch genRelateR Shiny Application}
\usage{
rungenRelateR()
}
\value{
Launches the Shiny application
}
\description{
Launches the interactive Shiny web application for genomic relatedness analysis
}
\examples{
\dontrun{
rungenRelateR()
}
}
