% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computationAndAnalysis.R
\name{validate_columns}
\alias{validate_columns}
\title{Helper function to validate data frame columns}
\usage{
validate_columns(df, required_cols, df_name)
}
\arguments{
\item{df}{Data frame to validate}

\item{required_cols}{Expected columns in the frame}

\item{df_name}{Name of the data frame for error message}
}
\description{
Helper function to validate data frame columns
}
